change "0" to Enter
