--- src/main/java/com/javarush/afanasenko/CryptoException.java
@@ -1,10 +1,0 @@
-package com.javarush.afanasenko;
-
-public class CryptoException extends RuntimeException {
-    public CryptoException(){
-    }
-    public CryptoException(String message){
-        super(message);
-    }
-
-}
