+++ src/main/java/com/javarush/afanasenko/ProjectRunner.java
@@ -1,48 +1,52 @@
 package com.javarush.afanasenko;
 
-//Тебе нужно разобраться как работает шифр Цезаря
+import com.javarush.afanasenko.action.BruteForse;
+import com.javarush.afanasenko.exception.CryptoException;
+import com.javarush.afanasenko.objects.Encoder;
+import com.javarush.afanasenko.objects.PathBuilder;
+import com.javarush.afanasenko.objects.TextHandler;
 
 import java.util.Scanner;
 
 public class ProjectRunner {
     public static void main(String[] args) {
-     while(true) {
-         printMenu();
-         Scanner sc = new Scanner(System.in);
-         String choice = sc.nextLine();
-         switch (choice) {
-             case "1":
-                 try {
-                     Encoder encoder = createEncoder(sc, "text\\text.txt");
-                     encoder.encode("encode","text\\encodedFile.txt", true);
-                     System.out.println("Файл успешно закодирован.");
-                 } catch (CryptoException ex) {
-                     System.out.println("\u001B[31m" + ex.getMessage());
-                 }
-                 break;
-             case "2":
-                 try {
-                     Encoder encoder = createEncoder(sc, "text\\encodedFile.txt");
-                     encoder.encode("decode", "text\\decodedFile.txt", true);
-                     System.out.println("Файл успешно декодирован.");
-                 } catch (CryptoException ex) {
-                     System.out.println("\u001B[31m" + ex.getMessage());
-                 }
-                 break;
-             case "3":
-                 try {
-                     BruteForse bruteForse = new BruteForse(createEncoder(sc, "text\\encodedFile.txt"));
-                     bruteForse.hack();
-                 }catch(CryptoException ex){
-                     System.out.println("\u001B[31m" + ex.getMessage());
-                 }
-                 break;
-             case "4":
-                 return;
-             default:
-                 System.out.println("\u001B[31m" + "Такой символ не поддерживается программой");
-         }
-     }
+        while (true) {
+            printMenu();
+            Scanner sc = new Scanner(System.in);
+            String choice = sc.nextLine();
+            switch (choice) {
+                case "1":
+                    try {
+                        Encoder encoder = createEncoder(sc, "text\\text.txt");
+                        encoder.encode("encode", "text\\encodedFile.txt", true);
+                        System.out.println("Файл успешно закодирован.");
+                    } catch (CryptoException ex) {
+                        System.out.println("\u001B[31m" + ex.getMessage());
+                    }
+                    break;
+                case "2":
+                    try {
+                        Encoder encoder = createEncoder(sc, "text\\encodedFile.txt");
+                        encoder.encode("decode", "text\\decodedFile.txt", true);
+                        System.out.println("Файл успешно декодирован.");
+                    } catch (CryptoException ex) {
+                        System.out.println("\u001B[31m" + ex.getMessage());
+                    }
+                    break;
+                case "3":
+                    try {
+                        BruteForse bruteForse = new BruteForse(createEncoder(sc, "text\\encodedFile.txt"));
+                        bruteForse.hack();
+                    } catch (CryptoException ex) {
+                        System.out.println("\u001B[31m" + ex.getMessage());
+                    }
+                    break;
+                case "4":
+                    return;
+                default:
+                    System.out.println("\u001B[31m" + "Такой символ не поддерживается программой");
+            }
+        }
     }
 
     public static void printMenu() {
@@ -54,13 +58,13 @@ public class ProjectRunner {
         System.out.println("4 - выход из программы");
     }
 
-    public static Encoder createEncoder(Scanner sc, String defaultPath) throws CryptoException{
-        System.out.println("Введите путь к файлу или используйте путь по умолчанию(введите 0)");
+    public static Encoder createEncoder(Scanner sc, String defaultPath) throws CryptoException {
+        System.out.println("Введите путь к файлу или используйте путь по умолчанию(нажмите enter)");
         String pathString = sc.nextLine();
         System.out.println("Введите ключ кодировки(для режима взлома нажмите любую цифру)");
         int key = sc.nextInt();
-        if(pathString.equals("0")) pathString = defaultPath;
+        if (pathString=="")pathString = defaultPath;
         PathBuilder builder = new PathBuilder(pathString);
-        return new Encoder((new TextHandler(builder.getPath())),key);
+        return new Encoder((new TextHandler(builder.getPath())), key);
     }
 }
