+++ src/main/java/com/javarush/afanasenko/exception/CryptoException.java
@@ -1,0 +1,10 @@
+package com.javarush.afanasenko.exception;
+
+public class CryptoException extends RuntimeException {
+    public CryptoException() {
+    }
+
+    public CryptoException(String message) {
+        super(message);
+    }
+}
